Reorganize infrastructure configs from global/config to infra package
